project/sync: add more project client methods to sync/editor/generic
