--- src/packages/project/types.ts
@@ -1,6 +1,0 @@
-import type { Client } from "./client";
-
-export interface SageSessionOpts {
-  client: Client;
-  path: string; // the path to the *worksheet* file
-}
