+++ src/packages/sync/editor/string/test/client-test.ts
@@ -15,6 +15,7 @@ import {
   FileWatcher as FileWatcher0,
 } from "../../generic/types";
 import { SyncTable } from "@cocalc/sync/table/synctable";
+import { ExecuteCodeOptionsWithCallback } from "@cocalc/util/types/execute-code";
 
 export class FileWatcher extends EventEmitter implements FileWatcher0 {
   private path: string;
@@ -169,5 +170,13 @@ export class Client extends EventEmitter implements Client0 {
   // account_id or project_id
   public client_id(): string {
     return this._client_id;
+  }
+
+  public sage_session({ path }): void {
+    console.log(`sage_session: path=${path}`);
+  }
+
+  public shell(opts: ExecuteCodeOptionsWithCallback): void {
+    console.log(`shell: opts=${JSON.stringify(opts)}`);
   }
 }
